updates to documentaiton

+++ CGAT.py
@@ -129,6 +129,9 @@ def publish_report( prefix = "",
     def _copy( src, dest ):
         dest = os.path.abspath( os.path.join( PARAMS["web_dir"], dest ) )
         if os.path.exists( dest ): shutil.rmtree( dest )
+        if not os.path.exists(src):
+            E.warn( "%s does not exist - skipped" % src )
+            return
         shutil.copytree( os.path.abspath(src), dest ) 
 
     # publish export dir via symlinking
